LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/allow-scripts-flag-changing-1-expected.txt

CONSOLE MESSAGE: Unable to do meta refresh due to sandboxing

PASS Meta refresh is blocked by the allow-scripts sandbox flag at its creation time, not when refresh comes due
